refactor: DevicePreivew 위치 변경
